Moved theme to its own file.
