+++ App.tsx
@@ -6,39 +6,11 @@ import { NavigationContainer } from '@react-navigation/native';
 import { RootNavigator } from './src/components/RootNavigator';
 import { Provider as StorageProvider } from 'storage/context';
 
-// Typescript custom theme declaration
-declare global {
-  namespace ReactNativePaper {
-    interface ThemeColors {
-
-    }
-    interface Theme {
-      iconSizes: {
-        small: number,
-        medium: number,
-        large: number
-      }
-    }
-  }
-}
-
-// Custom theme
-const customTheme = {
-  ...DefaultTheme,
-  colors: {
-    ...DefaultTheme.colors,
-
-  },
-  iconSizes: {
-    small: 30,
-    medium: 40,
-    large: 60
-  }
-};
+import { CustomTheme } from './src/CustomTheme';
 
 export default function App() {
   return (
-    <PaperProvider theme={customTheme}>
+    <PaperProvider theme={CustomTheme}>
       <NavigationContainer>
         <StorageProvider>
           <RootNavigator />
